Added "property" support.  Asset type called GSTProperty, having name, description, value attributes, can now be loaded into memory and accessed by name inside a JSP via EL.  Useful for constants, hard-coded values, etc. that need to be accessed in a web page.
Installation is not yet in place. (needs a flex asset, flex definition, and an attribute called "value").

Translation support should be added at some point (maybe?) and properties of type asset do not currently track any sort of dependency.  (Not sure if this is a problem or not).  

NOTE: This currently loads a cache dependency for all properties any time a single dependency is accessed.  This needs to be fixed.

Tag support is basic for now.

All tags tested by me.