Added a handy tag to do translation lookups based on a dimensionset after commercecontext.getrecommendation tag calls.  A companion tag that takes a single asset as input is probably warranted.  Typical syntax for this new tag is:
								<commercecontext:getrecommendations collectionid='${tag.id}' listvarname="reclist"/>
								<gsf:translate-list inlist="reclist" dimset="EEGlobal" />

The tag is smart and forgiving, and trivial to implement. For instance, outlist is optional (replacing inlist).  dimset can be either a dimensionset name or id.  locale is optional, an can be either a dimension id, a locale name, or, if not set, is auto-detected in ics as a variable, session variable, or the accept-language header.  Detailed logging explains what's going on.

Added required support infrastructure to LocaleUtils and DimensionUtils.